+++ docs/versionhistory.rst
@@ -18,6 +18,7 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 - Added ``stdin`` argument to ``anyio.run_process()`` akin to what
   ``anyio.open_process()``, ``asyncio.create_subprocess_…()``, ``trio.run_process()``,
   and ``subprocess.run()`` already accept (PR by @jmehnle)
+- Added the ``info`` property to ``anyio.Path`` on Python 3.14
 - Fixed traceback formatting growing quadratically with level of ``TaskGroup``
   nesting on asyncio due to exception chaining when raising ``ExceptionGroups``
   in ``TaskGroup.__aexit__``